Add class for group byte_busters Removed my class
